create 'multi_interval' submodule
