+++ kipoiseq/extractors/__init__.py
@@ -4,4 +4,5 @@ from .gtf import *
 from .vcf import *
 from .vcf_seq import *
 from .vcf_matching import *
+from .multi_interval import *
 from .protein import *
